PR113: style changes, moved some generally useful geometry helpers to GeometryMath.js
