+++ lib/bakeAmbientOcclusion.js
@@ -14,16 +14,17 @@ var Ray = Cesium.Ray;
 var ShaderSource = Cesium.ShaderSource;
 var WebGLConstants = Cesium.WebGLConstants;
 
+var byteLengthForComponentType = require('./byteLengthForComponentType');
 var clone = require('clone');
 var Jimp = require('jimp');
 var NodeHelpers = require('./NodeHelpers');
 var readAccessor = require('./readAccessor');
+var GeometryMath = require('./GeometryMath');
 
 module.exports = {
     bakeAmbientOcclusion: bakeAmbientOcclusion,
     generateRaytracerScene: generateRaytracerScene,
     computeAmbientOcclusionAt: computeAmbientOcclusionAt,
-    flattenTriangle: flattenTriangle,
     raytraceAtTriangleCenters: raytraceAtTriangleCenters,
     raytraceOverTriangleSamples: raytraceOverTriangleSamples
 };
@@ -151,9 +152,10 @@ function bakeToVertices(gltf, scene, raytracerScene, options) {
     NodeHelpers.forEachPrimitiveInScene(gltf, scene, checkShadingChain, parameters);
 
     // Clone materials, techniques, programs, and shaders as needed by primitives in other scenes.
-    for (var otherSceneID in gltf.scenes) {
-        if (gltf.scenes.hasOwnProperty(otherSceneID)) {
-            var otherScene = gltf.scenes[otherSceneID];
+    var sceneIDs = gltf.scenes;
+    for (var otherSceneID in sceneIDs) {
+        if (sceneIDs.hasOwnProperty(otherSceneID)) {
+            var otherScene = sceneIDs[otherSceneID];
             NodeHelpers.forEachPrimitiveInScene(gltf, otherScene, cloneShadingChain, parameters);
         }
     }
@@ -163,19 +165,21 @@ function bakeToVertices(gltf, scene, raytracerScene, options) {
     addAoToShaders(gltf, Object.keys(parameters.techniqueCloneIDs), options);
 
     // Edit the programs: add the ao attribute
-    for (var programID in parameters.programCloneIDs) {
-        if (parameters.programCloneIDs.hasOwnProperty(programID)) {
+    var programCloneIds = parameters.programCloneIDs;
+    for (var programID in programCloneIds) {
+        if (programCloneIds.hasOwnProperty(programID)) {
             gltf.programs[programID].attributes.push('a_ambientOcclusion');
         }
     }
 
+    var techniqueCloneIds = parameters.techniqueCloneIDs;
     // Edit the techniques: add ao to attributes, add ao to parameters
-    for (var techniqueID in parameters.techniqueCloneIDs) {
-        if (parameters.techniqueCloneIDs.hasOwnProperty(techniqueID)) {
+    for (var techniqueID in techniqueCloneIds) {
+        if (techniqueCloneIds.hasOwnProperty(techniqueID)) {
             var technique = gltf.techniques[techniqueID];
             technique.attributes.a_ambientOcclusion = 'vertex_ao';
             technique.parameters.vertex_ao = {
-                semantic: 'VERTEX_AO',
+                semantic: '_OCCLUSION',
                 type: WebGLConstants.FLOAT
             };
         }
@@ -219,7 +223,7 @@ function addVertexData(gltf, scene, raytracerScene) {
 
     gltf.buffers.aoBuffer = {
         input: {
-            byteLength : allAODataLength * 4, // (FLOAT). TODO: add option to bake to shorts
+            byteLength : allAODataLength * byteLengthForComponentType(WebGLConstants.FLOAT),
             type: 'arrayBuffer',
             uri: 'data:'
         },
@@ -245,7 +249,7 @@ function addVertexData(gltf, scene, raytracerScene) {
         var primitive = primitiveOrder[i];
         var primitiveVertexCount = parameters.primitiveVertexCounts[i];
         var name = 'accessor_' + parameters.primitiveNames[i] + '_AO';
-        primitive.attributes.VERTEX_AO = name;
+        primitive.attributes._OCCLUSION = name;
         gltf.accessors[name] = {
             bufferView: 'aoBufferView',
             byteOffset: byteOffset,
@@ -724,8 +728,8 @@ function trianglePixelMarch(uv0, uv1, uv2, position0, position1, position2, norm
             // Use this barycentric coordinate to compute the local space position and normal on the triangle
             var position = worldPositionScratch;
             var normal = worldNormalScratch;
-            sumBarycentric(barycentric, position0, position1, position2, position);
-            sumBarycentric(barycentric, normal0, normal1, normal2, normal);
+            GeometryMath.sumBarycentric(barycentric, position0, position1, position2, position);
+            GeometryMath.sumBarycentric(barycentric, normal0, normal1, normal2, normal);
 
             // Transform to world space
             Matrix4.multiplyByPoint(transform, position, position);
@@ -824,8 +828,8 @@ function raytraceAtTriangleCenters(primitive, meshPrimitiveID, parameters, node)
         barycentric.y = 1/3;
         barycentric.z = 1/3;
 
-        sumBarycentric(barycentric, positions[index0], positions[index1], positions[index2], position);
-        sumBarycentric(barycentric, normals[index0], normals[index1], normals[index2], normal);
+        GeometryMath.sumBarycentric(barycentric, positions[index0], positions[index1], positions[index2], position);
+        GeometryMath.sumBarycentric(barycentric, normals[index0], normals[index1], normals[index2], normal);
 
         // Transform to world space
         Matrix4.multiplyByPoint(transform, position, position);
@@ -885,7 +889,7 @@ function raytraceOverTriangleSamples(primitive, meshPrimitiveID, parameters, nod
         var position2 = positions[i2];
 
         // compute UVs for the triangle
-        flattenTriangle([position0, position1, position2], triangleUVs);
+        GeometryMath.flattenTriangle([position0, position1, position2], triangleUVs);
 
         var callbackParameters = {
             aoBuffer: aoBuffer,
@@ -964,7 +968,7 @@ function generateJitteredRay(position, normal, sampleNumber, sqrtNumberSamples) 
     randomDirection.y = t;
     randomDirection.z = v * Math.sin(u);
 
-    // Orient with texel's normal in world space
+    // Orient with given normal
     var theta = Math.acos(normal.y); // dot product of normal with y-up is normal.y
     var axis = Cartesian3.cross(randomDirection, normal, axisScratch);
     var rotation = Quaternion.fromAxisAngle(axis, theta, quaternionScratch);
@@ -1058,58 +1062,4 @@ function rayTriangle(ray, p0, p1, p2, cullBackFaces) {
     }
 
     return t;
-}
-
-var sumBarycentricScratch = new Cartesian3();
-
-function sumBarycentric(barycentric, vector0, vector1, vector2, result) {
-    result.x = 0.0;
-    result.y = 0.0;
-    result.z = 0.0;
-    Cartesian3.multiplyByScalar(vector0, barycentric.x, sumBarycentricScratch);
-    Cartesian3.add(result, sumBarycentricScratch, result);
-    Cartesian3.multiplyByScalar(vector1, barycentric.y, sumBarycentricScratch);
-    Cartesian3.add(result, sumBarycentricScratch, result);
-    Cartesian3.multiplyByScalar(vector2, barycentric.z, sumBarycentricScratch);
-    Cartesian3.add(result, sumBarycentricScratch, result);
-    return result;
-}
-
-var xAxis = new Cartesian3();
-var zAxis = new Cartesian3();
-var yAxis = new Cartesian3();
-// Returns the point's x/y coordinates in the triangle's plane where position0 is the origin
-// and position1 - position0 is the "x" axis.
-function flattenTriangle(positions, results) {
-
-    // Compute x, y, z axes. Z is considered to be the plane normal.
-    xAxis = Cartesian3.subtract(positions[1], positions[0], xAxis);
-    yAxis = Cartesian3.subtract(positions[2], positions[0], yAxis);
-    xAxis = Cartesian3.normalize(xAxis, xAxis);
-    yAxis = Cartesian3.normalize(yAxis, yAxis);
-    zAxis = Cartesian3.cross(xAxis, yAxis, zAxis);
-    zAxis = Cartesian3.normalize(zAxis, zAxis);
-    yAxis = Cartesian3.cross(zAxis, xAxis, yAxis);
-    yAxis = Cartesian3.normalize(yAxis, yAxis);
-
-    // Compute x/y coordinates by getting distance to normalized x and y axes in 3D
-    results[0].x = 0.0;
-    results[0].y = 0.0;
-    results[1].x = pointLineDistanceParametric(positions[1], positions[0], xAxis);
-    results[1].y = pointLineDistanceParametric(positions[1], positions[0], yAxis);
-    results[2].x = pointLineDistanceParametric(positions[2], positions[0], xAxis);
-    results[2].y = pointLineDistanceParametric(positions[2], positions[0], yAxis);
-
-    return results;
-}
-
-// Returns the parametric position on the line closest to the given point.
-// Math reference: http://paulbourke.net/geometry/pointlineplane/
-var pointOnLineScratch = new Cartesian3();
-function pointLineDistanceParametric(point, linePosition, lineDirection) {
-    var p1 = linePosition;
-    var p2 = Cartesian3.add(p1, lineDirection, pointOnLineScratch);
-    var p3 = point;
-    var length2 = Cartesian3.magnitudeSquared(lineDirection);
-    return ((p3.x - p1.x) * (p2.x - p1.x) + (p3.y - p1.y) * (p2.y - p1.y) + (p3.z - p1.z) * (p2.z - p1.z)) / length2;
 }
